+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.9.8] - Unreleased
 
+### Fixed
+
+- Remove unnecessary restriction on the `l` parameter of `D.van_genuchten()`.
 
 ## [0.9.7] - 2020-02-11